+++ test/CMakeLists.txt
@@ -26,7 +26,6 @@ add_grail_executable(SRC testWidgets.cc LIBS grail)
 add_grail_executable(SRC ../src/csp/GenericXDLClient.cc LIBS grail)
 add_grail_executable(SRC testPopupMenu.cc LIBS grail)
 add_grail_executable(SRC testStats.cc LIBS grail)
-add_grail_executable(SRC test_stats.cc LIBS grail)
 #add_grail_executable(BINNAME testSolar SRC DrawNASAEphemerisSolarSystem2d.cc LIBS grail)
 
 if(${MPV_FOUND})
